+++ R/check_if_installed.R
@@ -10,9 +10,11 @@
 #'   packages are installed, `FALSE` otherwise, and does not stop or throw a
 #'   warning. If `quietly = TRUE`, argument `stop` is ignored. Use this argument
 #'   to internally check for package dependencies without stopping or warnings.
-#' @param minimum_version String, representing the minimum package version that
-#'   is required. If `NULL`, no check for minimum version is done. Note
-#'   that `minimum_version` only works when `package` is of length 1.
+#' @param prompt If `TRUE`, will prompt the user to install needed package(s).
+#'   Ignored if `quietly = TRUE`.
+#' @param minimum_version A character vector, representing the minimum package
+#'   version that is required for each package. Should be of same length as
+#'   `package`. If `NULL`, no check for minimum version is done.
 #' @param ... Currently ignored
 #'
 #' @return If `stop = TRUE`, and `package` is not yet installed, the
@@ -34,40 +36,70 @@ check_if_installed <- function(package,
                                stop = TRUE,
                                minimum_version = NULL,
                                quietly = FALSE,
+                               prompt = interactive(),
                                ...) {
-  # does it need to be displayed?
-  is_installed <- sapply(package, requireNamespace, quietly = TRUE)
-  message <- NULL
 
+  is_installed <- sapply(package, requireNamespace, quietly = TRUE)
+  what_is_wrong <- what_you_can_do <- NULL
+
+  ## Test
   if (!all(is_installed)) {
     # only keep not-installed packages
     package <- package[!is_installed]
 
-    # prepare the message
-    if (length(package) > 1) {
-      message <- format_message(
-        paste0("Packages ", paste(sprintf("'%s'", package), collapse = " and "), " are required ", reason, "."),
-        paste0("Please install them by running install.packages(", paste(sprintf("\"%s\"", package), collapse = ", "), ").")
-      )
-    } else {
-      message <- format_message(
-        paste0("Package '", package, "' is required ", reason, "."),
-        paste0("Please install it by running install.packages('", package, "').")
-      )
-    }
-  } else if (!is.null(minimum_version) && utils::packageVersion(package) < package_version(minimum_version)) {
-    # prepare the message
-    message <- format_message(
-      paste0("Package '", package, "' is installed, but package version '", minimum_version, "' is required ", reason, "."),
-      paste0("Please update the package by running install.packages('", package, "').")
-    )
-  }
+    what_is_wrong <- sprintf("Package%s %s required %s.",
+                             if (length(package) > 1L) "s" else "",
+                             paste(sprintf("'%s'", package), collapse = " and "),
+                             reason)
 
-  if (!quietly && !is.null(message)) {
-    if (stop) stop(message, call. = FALSE) else warning(message, call. = FALSE)
+    what_you_can_do <- sprintf("Please install %s by running install.packages(%s).",
+                               if (length(package) > 1L) "them" else "it",
+                               paste(sprintf("\"%s\"", package), collapse = ", "))
+
+  } else if (!is.null(minimum_version)) {
+    needs_update <- utils::packageVersion(package) < package_version(minimum_version)
+
+    if (any(needs_update)) {
+      # only keep not-up-to-date packages
+      package <- package[needs_update]
+      minimum_version <- minimum_version[needs_update]
+
+      what_is_wrong <- sprintf("Package%s %s %s installed, but package version%s %s %s required.",
+                               if (length(package) > 1L) "s" else "",
+                               paste(sprintf("'%s'", package), collapse = " and "),
+                               if (length(package) > 1L) "are" else "is",
+                               if (length(package) > 1L) "s" else "",
+                               paste(sprintf("'%s'", minimum_version), collapse = " and "),
+                               if (length(package) > 1L) "are" else "is")
+
+      what_you_can_do <- sprintf("Please update %s by running install.packages(%s).",
+                                 if (length(package) > 1L) "them" else "it",
+                                 paste(sprintf("\"%s\"", package), collapse = ", "))
+    }
   }
 
   class(is_installed) <- c("check_if_installed", class(is_installed))
+
+  ## What do?
+  if (!is.null(what_is_wrong) && !quietly) {
+    if (prompt) {
+      what_you_can_do <- sprintf("Would you like to %s %s? [y/n]",
+                                 if (grepl("update", what_you_can_do)) "update" else "install",
+                                 if (grepl("them", what_you_can_do)) "them" else "it")
+
+      print_color(format_message(what_is_wrong), "red")
+      should_install <- readline(format_message(what_you_can_do))
+
+      if (tolower(should_install) == "y") {
+        utils::install.packages(package, ...)
+        return(invisible(is_installed))
+      }
+    } else {
+      message <- format_message(what_is_wrong, what_you_can_do)
+      if (stop) stop(message, call. = FALSE) else warning(message, call. = FALSE)
+    }
+  }
+
   invisible(is_installed)
 }